Add files via upload
Here's what I've been working on. The code itself is a little messy and lacking in comments, and I have had trouble getting the backend functionality (most likely because I am missing out on some fundamental aspect of it) to integrate with the GUI functionality. Also I need to edit a few things in the GUI, but I wanted to focus on the functionality first. I made a class called C_to_G in CLI_to_GUI, which is supposed to handle most of the integration. The C_to_G class inherits from the basicCLI class, and is passed around the various GUI's. Mostly what it can do is find address books that are already made, save new addressbooks, and display information-addresses and address information-that are already in the addressbooks.  It does not add or delete contacts, or update existing contacts.